+++ src/main/java/com/kakao_tech/community/exception/code/AuthErrorCode.java
@@ -1,10 +1,12 @@
-package com.kakao_tech.community.exception;
+package com.kakao_tech.community.exception.code;
+
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
 import org.springframework.http.HttpStatus;
 
-import lombok.AllArgsConstructor;
+import com.kakao_tech.community.exception.common.ErrorCode;
 
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
+import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
 public enum AuthErrorCode implements ErrorCode {
@@ -33,9 +35,9 @@ public enum AuthErrorCode implements ErrorCode {
     DIFFERENT_PASSWORD_CONFIRM(BAD_REQUEST, "DIFFERENT_PASSWORD_CONFIRM","입력된 비밀번호와 확인 비밀번호가 달라요."),
     DIFFERENT_SIGN_INFO(BAD_REQUEST, "DIFFERENT_SIGN_INFO","입력된 정보가 올바르지 않아요.");
 
-    private HttpStatus httpStatus;
-    private String code;
-    private String message;
+    private final HttpStatus httpStatus;
+    private final String code;
+    private final String message;
 
     public HttpStatus getHttpStatus() {
         return this.httpStatus;
@@ -48,6 +50,4 @@ public enum AuthErrorCode implements ErrorCode {
     public String getMessage() {
         return this.message;
     }
-
-
 }
