+++ src/main/java/com/kakao_tech/community/exception/common/RestApiException.java
@@ -1,0 +1,14 @@
+package com.kakao_tech.community.exception.common;
+
+import lombok.Getter;
+
+@Getter
+public class RestApiException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+    public RestApiException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
